refactor: simplify validation logic
- Streamlined validation methods to enhance clarity and reduce redundancy.
- Removed obsolete or unnecessary validation steps for better maintainability.
- Improved the overall structure for easier future modifications.
- Break up the Validation Service
- Remove the unsupported feature for SDS40 toggle
- Remove validation for SDS40 unsupported
