+++ helm_deploy/values-alt-preprod.yaml
@@ -21,7 +21,6 @@ generic-service:
     MANAGE-OFFENCES_API_URL: https://manage-offences-api-preprod.hmpps.service.justice.gov.uk
     BOTUS_FEATURE_TOGGLE: "true"
     SDS_EARLY_RELEASE_FEATURE_TOGGLE: "true"
-    SDS_EARLY_RELEASE_UNSUPPORTED_FEATURE_TOGGLE: "false"
     SDS_EARLY_RELEASE_HINTS_FEATURE_TOGGLE: "true"
     SUPPORT_INACTIVE_SENTENCES_AND_ADJUSTMENTS: "true"
     TORERA_OFFENCE_TO_MANUAL_JOURNEY: "false"
