+++ helm_deploy/values-dev.yaml
@@ -22,7 +22,6 @@ generic-service:
     MANAGE-OFFENCES_API_URL: https://manage-offences-api-dev.hmpps.service.justice.gov.uk
     BOTUS_FEATURE_TOGGLE: "true"
     SDS_EARLY_RELEASE_FEATURE_TOGGLE: "true"
-    SDS_EARLY_RELEASE_UNSUPPORTED_FEATURE_TOGGLE: "false"
     SDS_EARLY_RELEASE_HINTS_FEATURE_TOGGLE: "true"
     SUPPORT_INACTIVE_SENTENCES_AND_ADJUSTMENTS: "false"
     TORERA_OFFENCE_TO_MANUAL_JOURNEY: "true"
