+++ src/main/resources/application.yml
@@ -144,7 +144,6 @@ domain-events-sns:
 feature-toggles:
   botus: ${BOTUS_FEATURE_TOGGLE:false}
   sds-early-release: ${SDS_EARLY_RELEASE_FEATURE_TOGGLE:false}
-  sds-early-release-unsupported: ${SDS_EARLY_RELEASE_UNSUPPORTED_FEATURE_TOGGLE:false}
   sds-early-release-hints: ${SDS_EARLY_RELEASE_HINTS_FEATURE_TOGGLE:false}
   support-inactive-sentences-and-adjustments: ${SUPPORT_INACTIVE_SENTENCES_AND_ADJUSTMENTS:false}
   torera-offence-to-manual-journey: ${TORERA_OFFENCE_TO_MANUAL_JOURNEY:false}
